Moved thumbnail into its own package
