+++ internal/resolvers/default/resolver.go
@@ -19,6 +19,7 @@ import (
 	"github.com/Chatterino/api/internal/resolvers/youtube"
 	"github.com/Chatterino/api/pkg/cache"
 	"github.com/Chatterino/api/pkg/resolver"
+	"github.com/Chatterino/api/pkg/thumbnail"
 	"github.com/Chatterino/api/pkg/utils"
 	"github.com/go-chi/chi/v5"
 )
@@ -43,7 +44,8 @@ type R struct {
 
 	customResolvers []resolver.CustomURLManager
 
-	defaultResolverCache *cache.Cache
+	defaultResolverCache          *cache.Cache
+	defaultResolverThumbnailCache *cache.Cache
 }
 
 func (dr *R) HandleRequest(w http.ResponseWriter, r *http.Request) {
@@ -65,12 +67,31 @@ func (dr *R) HandleRequest(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
+func (dr *R) HandleThumbnailRequest(w http.ResponseWriter, r *http.Request) {
+	url, err := utils.UnescapeURLArgument(r, "url")
+	if err != nil {
+		_, err = w.Write(resolver.InvalidURL)
+		if err != nil {
+			log.Println("Error writing thumbnail response:", err)
+		}
+		return
+	}
+
+	response := dr.defaultResolverThumbnailCache.Get(url, r)
+
+	_, err = w.Write(response.([]byte))
+	if err != nil {
+		log.Println("Error writing thumbnail response:", err)
+	}
+}
+
 func New(baseURL string) *R {
 	r := &R{
 		baseURL: baseURL,
 	}
 
-	r.defaultResolverCache = cache.New("linkResolver", r.load, time.Duration(10)*time.Minute)
+	r.defaultResolverCache = cache.New("linkResolver", r.load, 10*time.Minute)
+	r.defaultResolverThumbnailCache = cache.New("thumbnail", thumbnail.DoThumbnailRequest, 10*time.Minute)
 
 	// Register Link Resolvers from internal/resolvers/
 	r.customResolvers = append(r.customResolvers, betterttv.New()...)
@@ -92,4 +113,5 @@ func Initialize(router *chi.Mux, baseURL string) {
 	defaultLinkResolver := New(baseURL)
 
 	router.Get("/link_resolver/{url}", defaultLinkResolver.HandleRequest)
+	router.Get("/thumbnail/{url}", defaultLinkResolver.HandleThumbnailRequest)
 }
