+++ pkg/thumbnail/thumbnail.go
@@ -1,5 +1,4 @@
-// TODO: this should potentially be split out into its own file
-package defaultresolver
+package thumbnail
 
 import (
 	"bytes"
@@ -19,7 +18,6 @@ import (
 	"github.com/Chatterino/api/pkg/cache"
 	"github.com/Chatterino/api/pkg/resolver"
 	"github.com/Chatterino/api/pkg/utils"
-	"github.com/go-chi/chi/v5"
 	"github.com/nfnt/resize"
 )
 
@@ -55,7 +53,7 @@ func buildStaticThumbnailByteArray(inputBuf []byte, resp *http.Response) ([]byte
 	return buffer.Bytes(), nil
 }
 
-func doThumbnailRequest(urlString string, r *http.Request) (interface{}, time.Duration, error) {
+func DoThumbnailRequest(urlString string, r *http.Request) (interface{}, time.Duration, error) {
 	url, err := url.Parse(urlString)
 	if err != nil {
 		return resolver.InvalidURL, cache.NoSpecialDur, nil
@@ -90,7 +88,7 @@ func doThumbnailRequest(urlString string, r *http.Request) (interface{}, time.Du
 		return resolver.NoLinkInfoFound, cache.NoSpecialDur, nil
 	}
 
-	if !isSupportedThumbnail(resp.Header.Get("content-type")) {
+	if !IsSupportedThumbnail(resp.Header.Get("content-type")) {
 		return resolver.NoLinkInfoFound, cache.NoSpecialDur, nil
 	}
 
@@ -113,7 +111,7 @@ func doThumbnailRequest(urlString string, r *http.Request) (interface{}, time.Du
 	return image, 10 * time.Minute, nil
 }
 
-func isSupportedThumbnail(contentType string) bool {
+func IsSupportedThumbnail(contentType string) bool {
 	for _, supportedType := range supportedThumbnails {
 		if contentType == supportedType {
 			return true
@@ -121,30 +119,4 @@ func isSupportedThumbnail(contentType string) bool {
 	}
 
 	return false
-}
-
-func thumbnail(w http.ResponseWriter, r *http.Request) {
-	url, err := utils.UnescapeURLArgument(r, "url")
-	if err != nil {
-		_, err = w.Write(resolver.InvalidURL)
-		if err != nil {
-			log.Println("Error writing response:", err)
-		}
-		return
-	}
-
-	response := thumbnailCache.Get(url, r)
-
-	_, err = w.Write(response.([]byte))
-	if err != nil {
-		log.Println("Error writing response:", err)
-	}
-}
-
-var (
-	thumbnailCache = cache.New("thumbnail", doThumbnailRequest, 10*time.Minute)
-)
-
-func InitializeThumbnail(router *chi.Mux) {
-	router.Get("/thumbnail/{url}", thumbnail)
 }
